refactor: member 도메인 패키지 생성
member 도메인 패키지 생성에 따라 기존 auth에 속해있던 member관련 코드를 이동한다.
